refactor

+++ firmware/arduino_tb6612fng_25GA_370/arduino_tb6612fng_25GA_370.ino
@@ -17,7 +17,7 @@ Motor m1(ENC_COUNT_REV,
          3);
 
 //bool stopTune = false;
-unsigned int command;
+byte command;
 double velocity;
 #define MOVE_MOTOR_1 1
 #define MOVE_MOTOR_2 2
